move into separate xyz mod
