+++ src/cliffs.rs
@@ -9,7 +9,7 @@ use std::io::{BufWriter, Write};
 use std::path::Path;
 
 use crate::config::Config;
-use crate::io::XyzInternalReader;
+use crate::io::xyz::XyzInternalReader;
 use crate::vec2d::Vec2D;
 
 pub fn makecliffs(config: &Config, tmpfolder: &Path) -> Result<(), Box<dyn Error>> {
